More return info on LtiOutcomesHelper methods
For one reason or another, we've been running into errors when using the
LtiOutcomesHelper, and it's
difficult to diagnose the problem when the reason is hidden. This should
expose a bit more info on why a false result is being sent back. I added
a new BasicResult class that can be implicitly cast to a bool. The root
and v1 LisResult classes now implement this BaiscResult and use its
IsValid property. For the message, Exceptions return back its ToString
result, and imsx_StatusInfoType will return the imsx_description.