https://github.com/andyfmiller/LtiLibrary.Owin/issues/1
Signature check in provider does not include querystring parameters in signature base (but consumer does).

To fix this, there are now two versions of OAuthRequest.GenerateSignature. One-to be used by consumers-takes a url, parameters, and consumersecret. This allows the consumer to perform the parameter substitution before signing the request. The other uses the current value. This allows the provider to simply parse the request and calculate the signature.

Also add support for custom parameter substitution in querystring. For example, ?custom_username=$User.username will be transformed into ?custom_username=someusername.