In HttpResponseExtensions, encode the parameters
The value of a parameter can contain characters that break the HTML form post.
E.g. a single quote in lis_person_name_full = "d'Artagnan", would cause a form post with input value="d".
NB I tested my fix on an older version of the LtiLibrary, so I did not actually compile and test my proposed change - sorry for that.